requirements.txt
streamlit
PyMuPDF
pyzbar
Pillow
requests
pandas